+++ Sources/DeckKit/Favorites/Favoritable.swift
@@ -1,0 +1,12 @@
+//
+//  Favoritable.swift
+//  DeckKit
+//
+//  Created by Daniel Saidi on 2023-06-12.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+/**
+ This protocol can be implemented by anything that can be markes as a favorite.
+ */
+public protocol Favoritable: Identifiable {}