+++ Sources/DeckKit/Favorites/UserDefaultsFavoriteService.swift
@@ -3,7 +3,7 @@
 //  DeckKit
 //
 //  Created by Daniel Saidi on 2020-09-17.
-//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//  Copyright © 2020-2023 Daniel Saidi. All rights reserved.
 //
 
 import Foundation